fix efficiency crash; change efficiency to be 0% at border tiles; add proper background to room decorations;

+++ collective.h
@@ -367,7 +367,7 @@ class Collective : public CreatureView, public EventListener, public Task::Callb
   map<UniqueId, MarkovChain<MinionTask>> SERIAL(minionTasks);
   map<UniqueId, string> SERIAL(minionTaskStrings);
   map<SquareType, set<Vec2>> SERIAL(mySquares);
-  map<Vec2, double> SERIAL(squareEfficiency);
+  map<Vec2, int> SERIAL(squareEfficiency);
   void updateEfficiency(Vec2, SquareType);
   double getEfficiency(Vec2) const;
   bool hasEfficiency(Vec2) const;
